+++ src/displays/DisplayST7735Adapter.h
@@ -24,30 +24,30 @@ public:
     LOGGER_JUSTINFO("init: w:%d, h:%d, r:%d", conf->width, conf->height, conf->rotation);
     LOGGER_JUSTINFO("  pins: l:%d, r:%d", conf->lightPin, conf->resetPin);
     LOGGER_JUSTINFO("   i2c: adr:%d, sda:%d, scl:%d", conf->i2cAddress, conf->i2cSDA, conf->i2cSCL);
-    LOGGER_JUSTINFO("   spi: cs:%d, dc:%d, mosi:%d, miso:%d, clk:%d", conf->spiCS, conf->spiDC, conf->spiMOSI, conf->spiMISO, conf->spiCLK);
+    LOGGER_JUSTINFO("   spi: cs:%d, dc:%d, mosi:%d, miso:%d, clk:%d", conf->csPin, conf->dcPin, conf->spiMOSI, conf->spiMISO, conf->spiCLK);
 
     if ((conf->busmode == BUSMODE_ANY) || (conf->busmode == BUSMODE_SPI)) {
 
 #if defined(ESP32)
 
 #if defined(CONFIG_IDF_TARGET_ESP32C3)
-      bus = new Arduino_HWSPI(conf->spiDC, conf->spiCS, conf->spiCLK, conf->spiMOSI, conf->spiMISO);
+      bus = new Arduino_HWSPI(conf->dcPin, conf->csPin, conf->spiCLK, conf->spiMOSI, conf->spiMISO);
 
 #elif defined(CONFIG_IDF_TARGET_ESP32S3)
       bus = new Arduino_ESP32SPI(
-        conf->spiDC,
-        conf->spiCS,
+        conf->dcPin,
+        conf->csPin,
         conf->spiCLK,
         conf->spiMOSI,
         conf->spiMISO,
         HSPI /* spi_num */
       );
 #else
-      bus = new Arduino_HWSPI(conf->spiDC, conf->spiCS, conf->spiCLK, conf->spiMOSI, conf->spiMISO);
+      bus = new Arduino_HWSPI(conf->dcPin, conf->csPin, conf->spiCLK, conf->spiMOSI, conf->spiMISO);
 #endif
 
 #elif defined(ESP8266)
-      bus = new Arduino_HWSPI(conf->spiDC, conf->spiCS);
+      bus = new Arduino_HWSPI(conf->dcPin, conf->csPin);
 #endif
     } else {
       // any bus but SPI is not implemented