+++ src/displays/DisplayST7789Element.h
@@ -39,9 +39,13 @@ public:
    */
   static bool registered;
 
-  /**
-   * @brief Activate the Element.
-   */
+
+  /// @brief Setup the element so it can be started and stopped.
+  virtual void setup() override;
+
+
+  /// @brief Activate the Element and register a Display Adapter
+  /// in the board.
   virtual void start() override;
 };